Removed log warning on median filter
The median filter "nan" warning on jwst/coron/median_replace_img.py is worthless
for MIRI coronagraphs because the subarrays have large regions of non-science
pixels that get flagged DO_NOT_USE, resulting in tens of thousands of warning
messages per integration and a huge slowdown. Since there are such large regions
of DO_NOT_USE pixels, this warning is not informative, so let's get rid of it.